Refactored the use of export_to_json and removed RemoteUtils.py
