+++ src/refresh_remote_catalog.py
@@ -17,7 +17,6 @@ from git import Repo
 import yaml
 from os.path import dirname, exists
 from json import load
-import RemoteUtils
 import Utils
 
 def pull_from_remote(path_to_repo, repo_name):
@@ -60,8 +59,8 @@ def update_json_file(repo_name):
         yaml_file = data['yaml-file']
         
     if yaml_file.endswith('.yaml'):
-        df, population, filter, sortables, display, site_details = RemoteUtils.parse_yaml_to_json(CATALOGS_PATH, CATALOGS_PATH + '/' + repo_name + '/' + yaml_file)
-        RemoteUtils.export_to_json(CATALOGS_PATH, df, population, filter, sortables, display, site_details, repo_name, yaml_file)
+        df, population, filter, sortables, display, site_details = Utils.parse_yaml_to_json(CATALOGS_PATH, CATALOGS_PATH + '/' + repo_name + '/' + yaml_file)
+        Utils.export_to_json(CATALOGS_PATH, df, population, filter, sortables, display, site_details, repo_name, yaml_file)
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(
